changelog ID as a sequence
As noted elsewhere, auto-increment sequences in postgres, mysql, and
cockroach can have gaps: when transactions that increment a number are
rolled back, a "gap" is left.

I didn't realize this, and it's problematic for the external changelog
API. I'll need to re-write the changelog inserter to query for the
current highest number when inserting, which is unfortunate and might
slow down bulk imports.