remove 'IS NOT NULL' identifier index constraints
These seemed to be resulting in table scans on 404s in QA with postgres
10, despite the adding "IS NOT NULL" WHERE clauses earlier. Query time
was very significant, even for the creator table (1.2 seconds or so on
SSD).

I looked at using hash indices (which have improved in postgres 10),
which could save index size (disk and RAM) and potentially be faster for
these trivial exact lookups, but didn't go for it at this time.