don't require auth for editgroup annotations
Amazing that this bug found it's way through... because most testing is
from Python, and was having serious auth config leakage with
python_client.

We're still in 0.x, and this is such a small/eggregious bug that i'm not
going to tag as a backwards-incompatible schema update (but will note in
CHANGELOG).