start new ES container worker kafka group
The previous group seems to have gotten corrupted; my hypothesis is that
this is due to pykafka being somewhat flakey, and am planning to move to
librdkafka anyways. Re-indexing all the containers is pretty small/easy,
so starting a new consumer group works find in this case; release
indexer would be a bigger problem.