update rust README re: openssl
I believe an openssl library is still required locally, but with the
SSL/TLS removal it now doesn't matter if it is OpenSSL 1.0 or 1.1.

This is only a temporary work-around. When we update rust code
generation, we will need to revisit these changes. The current version
of swagger-rs still depends on HTTPS and OpenSSL 1.0 (via dependencies).