bump required rust to 1.36
This isn't a fatcat rust requirement, but instead a diesel requirement,
via rust-smallvec, which in v1.0 uses the alloc crate:

  https://github.com/servo/rust-smallvec/issues/73

I think the reason this came up now is that diesel-cli is an
application and doesn't have a Cargo.lock file, and the build was
updated. Using some binary mechanism to install these dependencies would
be more robust, but feels like a yak shave right now.