datacite: add conversion fixtures
The `test_datacite_conversions` function will compare an input
(datacite) document to an expected output (release entity as JSON). This
way, it should not be too hard to add more cases by adding: input,
output - and by increasing the counter in the range loop within the
test.

To view input and result side by side with vim, change into the test
directory and run:

    tests/files/datacite $ ./caseview.sh 18