fatcatd: fix corner-case in abstracts insertion
Both the release_abstract relation table and the abstracts table inserts
were being gated by a check on new abstracts table rows.

I *think* the chance of this having caused problems is low. The most
likely would have been updates to exiting entities that somehow removed
the abstracts content, but not the sha1 keys. This is not the default
behavior of the API: either the entire abstract (content and hash) is
returned, or the abstracts are hidden entirely (via `hide` flag).

Still, best to be careful!