refactor fatcat_import kafka group names
My current understanding is that consumer group names should be
one-to-one with topic names. I previously though offsets were stored on
a {topic, group} key, but they seem to be mixed and having too many
workers in the same group is bad. In particular, we don't want
cross-talk or load between QA and prod.

All these topics are caught up in prod, so deploying this change and
restarting workers should be safe.

This commit does not update the elasticsearch or entity updates workers.