catch ApiValueError in some generic API calls
The motivation for this change is to handle bogus revision IDs in URLs,
which were causing 500 errors not 400 errors. Eg:

    https://qa.fatcat.wiki/file/rev/5d5d5162-b676-4f0a-968f-e19dadeaf96e%2B2019-11-27%2B13:49:51%2B0%2B6

I have no idea where these URLs are actually coming from, but they
should be 4xx not 5xx.

Investigating made me realize there is a whole category of ApiValueError
exceptions we were not catching and should have been.