version bump fatcat_openapi_client
This is to push a new package version to pypi with the newer codegen.
The immediate motivation for this is to fix keyword errors with "async"
parameters conflicting with "async" as a keyword in Python 3.7.

I guess that means this is probably actually a breaking change to the
python API, though not to the API itself? Hrm.

All the other changes are typo corrections from the fatcat-openapi2.yml
spec document.