experimentally update semantic-ui to fomantic-ui 2.8.6
fomatic-ui is a "community fork" which has had more active recent
feature additions, though i'm not sure whether the quality has been as
good as the main semantic-ui project.

the proximal motivation to upgrade/switch is to get multi-bar progress
bars, to display overall preservation status in a single bar as a
fraction.