generic API error page
This error handler and view page currently works much better than the
"flash()" infrastructure built-in to flask, which uses cookies and
mostly does not work with our views and layouts.

Would like to gradually migrate almost all API errors in the web
interface to just raising errors that get rendered on an error page,
instead of calling `abort(ae.status)`.