web: route constraints on fcids and UUIDs
Instead of accepting any string for these parameters and throwing a 400
error if not the correct type, implement better route matching at the
framework level and return more 404s.

This resolves several outstanding sentry exceptions.

The "flask-uuid" was imported and seems to have been configured for this
purpose previously, but I guess I never finished configuring it.