api: add SQL 'ORDER BY' to many reads to stabilize API array ordering
The hope is to make things like file entity URLs, fileset manifests, and
other arrays in the JSON API "stable", meaning that if you create an
entity with a list of a given order, a read back (in any environment,
including prod/QA, bulk dumps, etc) will return the array with the same
sort order.

This was informally happening most of the time, but occasionally not (!)

Assumption is that these sorts will have little or no performance
impact, as the common case is less than a dozen elements, and the hard
cases are a few thousand at most, and there is already a sorted index.