add checksum for face wheel
one should not do this but the Pipfile is so fragile given its age I
can't regenerate the whole Pipfile.lock. Thus, this commit just adds a
checksum for the new face wheel that's on pypi (and is automatically
grabbed by pip).

I got both the source install we used to use and the new wheel and
compared them locally to ensure they were identical.