remove explicit transactions from migrations
With these in place I get an opaque "operation must be called from a
transaction" error when running diesel.

As best as I can figure, the newest diesel is automatically running
these migrations in a transaction (which I would expect it to do) and
introducing a second `COMMIT` which fails.

I could also be wrong and this is downstream of running postgres 15; I
didn't check.